basic routing between all views setup
