+++ src/components/About/About.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class About extends Component {
+    render() {
+        return (
+            <div>
+                
+            </div>
+        );
+    }
+}
+
+export default About;
