+++ src/components/TrailDetails/TrailDetails.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class TrailDetails extends Component {
+    render() {
+        return (
+            <div>
+                
+            </div>
+        );
+    }
+}
+
+export default TrailDetails;
