+++ src/components/TrailResults/TrailResults.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class TrailResults extends Component {
+    render() {
+        return (
+            <div>
+                
+            </div>
+        );
+    }
+}
+
+export default TrailResults;
