+++ src/components/landingPage/LandingPage.js
@@ -1,13 +1,29 @@
 import React, { Component } from 'react';
+import TrailSearch from '../../containers/trail-search';
+import { getTrailData } from '../../ducks/trail';
+import { connect } from 'react-redux';
 
 class LandingPage extends Component {
+    constructor(props) {
+  super(props);
+  this.state = {
+    trailData: {}
+  }
+}
     render() {
         return (
             <div>
-                LandingPage
+                <TrailSearch />
             </div>
         );
     }
 }
 
-export default LandingPage;
+function mapStateToProps(state) {
+    return {
+      info: state.trailReducer.trailData,
+      loading: state.trailReducer.loading
+    }
+  }
+
+export default connect(mapStateToProps, {getTrailData})(LandingPage);
